fix build break

+++ blitzwing-common/src/main/java/com/ebay/hadoop/blitzwing/memory/MemoryManager.java
@@ -1,5 +1,5 @@
 package com.ebay.hadoop.blitzwing.memory;
 
 public interface MemoryManager {
-  void freeBuffer(long memoryAddress, long length);
+  void freeBuffer(long memoryAddress, int length);
 }
